Error Reporting (#7)
* Add address error reporting for supporting bus protocols

AXI4 Lite and Wishbone both support the feature of returning an error
when an invalid address (i.e. an address belonging to the memory space
of a block but without associated register) is accessed. Previously,
such accesses where acknowledged without indication to the user that the
address is invalid.

This commit adds an option `bus-error` activating the report error
feature for the two bus protocols. Now in case of an addressing error,
- the AXI4 Lite response signal (`BRESP` for write or `RRESP` for read
  requests) is non-zero returning `SLVERR` (`0b10`)
- the Wishbone request is not acknowledged (`ACK` stays deasserted) but
  instead an error is returned (`ERR` is asserted).
Furthermore, with an active `bus-error` option, the AXI4 Lite
interface behaves now differently when put in reset: Its output signal
are configured to allow all handshakes while the two response signal
indicate an error. This allows to access the interface while being in
reset without stalling the bus.

If the `bus-error` is disabled (default), the exact same output is
generated as with the previous implementation.

For each of the two modified bus interfaces, the testbench package has
been updated to support the error reporting. Two testbenches have been
added, verifying the the error reporting behavior.