Move register access type out of documentation table
Although the original LaTeX generation implementation tried to match
the HTML generation, having limited horizontal space on a paper
document (compared to a screen) makes it a better choice to state the
access type only once. Furthermore, it has to be the same for all fields
of a register by definition anyway.