Update SV testfiles
It is unclear to me how the GitHub action jobs in #27 passed without
error. Also, I normally run all tests before pushing. Still, it seems
that some updated SV testfiles were likely be masked by a .gitignore
file and not committed in #27. Or, the combination of the two merge
requests #27 and #28 were not compatible.

This oversight should have been caught by the GitHub action job. As it
did not, here the remaining missing files.