Fix write mask application for read-write wires
Write masks were added in #16. Similar code between register fields and
wire fields was used. Since register fields are assigned in a sequential
process, using the same signal name on the right side of the assignment
(i.e., the current value) and on the left side (i.e. the new value) is
permitted. This is not valid for wires where such a code creates an
infinite loop. Instead, when using write masks with read-write wire
fields, the input wire value has to be used on the right side of the
assignment (i.e., the current value), while the output wire value has to
be used on the left side (i.e. the new value).