Do not OR autoclear fields with previous value
When using a write mask, fields are normally updated like following
new_data = (old_data and !write_mask) or (write_data and write_mask)
Since in the case of the autoclear field, the previous data is
irrelevant it should not be ORed. Otherwise, two consecutive write
requests to the same register with different masks might keep the field
asserted.