Align write strobe in case of register of type "wire"
The write strobe should be asserted synchronously with the update of
the register in case of a write request. Since the register has to first
sample the write data, the write strobe is a delay version of the write
request.
In case of a register of type "wire" this is no longer true: There is no
register and as such the write data is directly forwarded to the output
ports. Hence, the write strobe also should be forwarded without prior
sampling. This is fixed with this commit.