add busgroup support for axi4litebus
Based on existing code for wbbus, which has been moved and generalised
to BusOptions in busparam.py. Wbbus now use the same resize% functions
as axi4litebus. Axi4litebus uses Ibus' add_bus while wbbus doesn't,
which yields a difference in the generated hdltree which is the
motivation behind the print_verilog/vhdl and hdlopt exclusions of
HDLInterfaceSelect in this commit.

For AXI4-Lite, the new terms manager/subordinate are used for the
generated records. The package itself will be added to documentation
and the testfiles in the repo.